+++ konst/src/array.rs
@@ -15,8 +15,8 @@ macro_rules! leak_warning {
 }
 use leak_warning;
 
-/// Const equivalent of
-/// [`array::map`](https://doc.rust-lang.org/std/primitive.array.html#method.map).
+/// Superceeded by [`map_`],
+/// const version of [`array::map`](https://doc.rust-lang.org/std/primitive.array.html#method.map).
 ///
 #[doc = leak_warning!()]
 ///
@@ -83,16 +83,61 @@ use leak_warning;
 ///
 pub use konst_kernel::array_map as map;
 
-/// Const equivalent of [`array::from_fn`](core::array::from_fn).
+#[cfg(feature = "rust_1_83")]
+#[doc(hidden)]
+pub mod __array_macros_2;
+
+#[cfg(feature = "rust_1_83")]
+macro_rules! drop_warning {
+    () => {
+        concat!(
+            "# Note",
+            "\n\n",
+            "`return` inside the closure passed to this macro ",
+            "attempts to return from the function where this macro is called, ",
+            "which drops the array elements, ",
+            "and dropping isn't allowed in const as of Rust 1.83.0.",
+            "\n\n",
+            "The same applies to `?`, ",
+            "and labelled `break`/`continue` into labels from outside the closure.",
+        )
+    };
+}
+
+#[cfg(feature = "rust_1_83")]
+use drop_warning;
+
+/// Const equivalent of
+/// [`array::map`](https://doc.rust-lang.org/std/primitive.array.html#method.map)
+///
+#[doc = drop_warning!()]
+///
+/// # Example
+///
+/// ```rust
+/// assert_eq!(PAIRS, [(3, "hello"), (5, "world"), (8, "foo")]);
+///
+/// const PAIRS: [(u8, &str); 3] =
+///     swap_pairs([("hello", 3), ("world", 5), ("foo", 8)]);
+///
+/// const fn swap_pairs<T, U, const N: usize>(pairs: [(T, U); N]) -> [(U, T); N] {
+///     konst::array::map_!(pairs, |pair: (T, U)| {
+///         // need to use `destructure` to destructure types that may contain Drop fields
+///         konst::destructure!{(a, b) = pair}
+///         (b, a)
+///     })
+/// }
+/// ```
+///
+#[doc(inline)]
+#[cfg(feature = "rust_1_83")]
+#[cfg_attr(feature = "docsrs", doc(cfg(feature = "rust_1_83")))]
+pub use crate::__array_map_by_val as map_;
+
+/// Superceeded by [`from_fn_`], const version of
+/// [`array::from_fn`](core::array::from_fn).
 ///
 #[doc = leak_warning!()]
-///
-/// # Limitations
-///
-/// When the array type is annotated, the array type must be one of:
-/// - Square brackets (e.g: `from_fn!([usize; 10] => |i| i)`)
-/// - A parenthesized type (e.g: `from_fn!((foo::ArrayAlias) => |i| i * 2)`)
-/// - A single identifier (e.g: `from_fn!(ArrayAlias => |i| func(i))`)
 ///
 /// # Example
 ///
@@ -113,3 +158,30 @@ pub use konst_kernel::array_map as map;
 /// ```
 ///
 pub use konst_kernel::array_from_fn as from_fn;
+
+/// Const equivalent of [`array::from_fn`](core::array::from_fn).
+///
+#[doc = drop_warning!()]
+///
+///
+/// # Example
+///
+/// ```rust
+/// use konst::array;
+///
+/// {
+///     const POWERS: [u64; 5] = array::from_fn_!(|i| 2u64.pow(i as u32));
+///
+///     assert_eq!(POWERS, [1, 2, 4, 8, 16]);
+/// }
+///
+/// // Annotating the array type
+/// assert_eq!(
+///     array::from_fn_!([&str; 6] => |i| konst::string::str_up_to("hello", i)),
+///     ["", "h", "he", "hel", "hell", "hello"],
+/// );
+/// ```
+///
+#[cfg(feature = "rust_1_83")]
+#[cfg_attr(feature = "docsrs", doc(cfg(feature = "rust_1_83")))]
+pub use crate::__array_from_fn2 as from_fn_;