+++ konst/tests/misc_tests/array_macros2_ui_tests/non_local_return_map_err.rs
@@ -1,0 +1,7 @@
+#[allow(unreachable_code)]
+const _: [usize; 1] = 'out: {
+    konst::array::map_!([10], |x| break 'out [x * 2])
+};
+
+
+fn main() {}